Renames JsonReaders => BitcoindJsonReaders
