+++ rpc/src/main/scala/org/bitcoins/rpc/client/NodeRpc.scala
@@ -2,7 +2,7 @@ package org.bitcoins.rpc.client
 
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.rpc.jsonmodels.GetMemoryInfoResult
-import org.bitcoins.rpc.serializers.JsonReaders._
+import org.bitcoins.rpc.serializers.BitcoindJsonReaders._
 import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
 import play.api.libs.json.{ JsArray, JsString, Json }
