+++ rpc/src/main/scala/org/bitcoins/rpc/client/RpcOpts.scala
@@ -4,7 +4,7 @@ import org.bitcoins.core.crypto.{ DoubleSha256Digest, ECPrivateKey }
 import org.bitcoins.core.currency.Bitcoins
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.BitcoinAddress
-import org.bitcoins.rpc.serializers.BitcoindJsonWriters
+import org.bitcoins.rpc.serializers.BitcoindJsonWriters._
 import org.bitcoins.core.protocol.script.ScriptPubKey
 import play.api.libs.json.{ Json, Writes }
